cruft update, manual resolving of hunks
