+++ .github/workflows/publish.yml
@@ -8,6 +8,8 @@ env:
   python-version: 3.12
 
 jobs:
+  
+
   setup:
     name: Checkout and Install
     runs-on: ubuntu-latest
@@ -70,7 +72,10 @@ jobs:
 
   build:
     name: Build Package
-    needs: [compare-versions, test]
+    needs:
+      - compare-versions
+      - test
+      
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
@@ -81,6 +86,9 @@ jobs:
         uses: astral-sh/setup-uv@v5
         with:
           python-version: ${{ env.python-version }}
+
+      
+
       - name: Build package
         run: uv build
       - name: Upload dist artifacts
