+++ docs/learn/versioning.md
@@ -1,0 +1,9 @@
+# Note on versioning
+
+Keeping track of versions is essential for every software and semantic data modeling activity. We follow [PEP440](https://peps.python.org/pep-0440/) when reporting versions, fetching metadata at run time using [`importlib.metadata`](https://docs.python.org/3/library/importlib.metadata.html). Several versions need to be distinguished. The version of `pynxtools`, the one of the `NeXus definitions`, and those of `h5py` and the `HDF5` library. [See the respective section of the `pynxtools` documentation](https://fairmat-nfdi.github.io/pynxtools/learn/pynxtools/versioning.html) that explains where these versions can be found.
+
+## `pynxtools-plugin` version
+
+We consider it best practice that also each plugin of `pynxtools` writes its version as a `programID` instance of [`NXprogram`](https://github.com/FAIRmat-NFDI/nexus_definitions/blob/fairmat/base_classes/NXprogram.nxdl.xml). `ID` serves as a placeholder that can be either omitted or replaced by an integer if multiple programs need to be distinguished. Counting should start at one, i.e., `program` or `program1` are preferred options.
+Specifically, the field `program` inside the `programID` instance should store the distribution (package) name. The attribute `program/@version` should store a version string following [PEP440](https://peps.python.org/pep-0440/).
+
