+++ mkdocs.yaml
@@ -7,14 +7,16 @@ edit_uri: ""
 
 nav:
   - Home: index.md
-  - How-tos:
-    - how-tos/installation.md
-    - how-tos/convert_data.md
-    - how-tos/download_rod.md
   - Tutorials:
     - tutorial/installation.md
     - tutorial/contributing.md
     - tutorial/adjust_config_file.md
+  - How-tos:
+    - how-tos/installation.md
+    - how-tos/convert_data.md
+    - how-tos/download_rod.md
+  - Learning:
+    - learn/versioning.md
   - Reference:
     - reference/witec.md
     - reference/rod.md
@@ -75,5 +77,4 @@ extra_css:
 
 extra_javascript:
   - javascript.js
-  - https://polyfill.io/v3/polyfill.min.js?features=es6
   - https://cdn.jsdelivr.net/npm/mathjax@3/es5/tex-mml-chtml.js
