+++ tests/nomad/test_nomad_apps.py
@@ -15,15 +15,21 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+"""Tests for the NOMAD app."""
 
-import importlib.metadata
+import pytest
+
+try:
+    import nomad  # noqa: F401
+except ImportError:
+    pytest.skip(
+        "Skipping NOMAD app tests because nomad-lab is not installed",
+        allow_module_level=True,
+    )
 
 
-def get_pynxtools_raman_version() -> str:
-    """Attempt getting the version of pynxtools at runtime with fallback."""
-    # for a discussion whether to collect at build or runtime see
-    # https://discuss.python.org/t/please-make-package-version-go-away/58501
-    try:
-        return importlib.metadata.version("pynxtools-raman")
-    except importlib.metadata.PackageNotFoundError:
-        return f"unknown_version"
+def test_importing_app():
+    # this will raise an exception if pydantic model validation fails for the app
+    from pynxtools_raman.nomad.apps import raman_app  # noqa: PLC0415
+
+    assert raman_app.app.label == "Raman"
